requirements.txt
httpx==0.27.0
beautifulsoup4==4.12.3
lxml==5.2.1

sqlalchemy==2.0.30
psycopg2-binary==2.9.9

celery[redis]==5.4.0
redis==5.0.4

catboost==1.2.5
scikit-learn==1.5.0
pandas==2.2.2
joblib==1.4.2
numpy==1.26.4

python-telegram-bot==21.3

matplotlib==3.9.0

python-dotenv==1.0.1
shap==0.45.1

groq>=0.9.0

pytest==8.2.2
pytest-asyncio==0.23.7
